requirements.txt
django==4.2
djangorestframework
psycopg2-binary
python-dotenv
flask
requests
ollama
streamlit
